[WebAPI] Modify the test case according to the comments
Remove the un-referenced variable.
Remove the wrong test checkpoin.
Change the presentation.html directory under support
The async_test can't change to promise_test as the event handler
The onconnect is not triggered in this code, if combine with close/terminate, may miss this checkpoint. So design as 3 test cases.